Make TexturedCharacterGlyph public
