+++ osu.Framework/IO/Stores/FontStore.cs
@@ -7,7 +7,6 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using osu.Framework.Logging;
 using System.Collections.Concurrent;
-using System.Runtime.CompilerServices;
 using osu.Framework.Platform;
 using osu.Framework.Text;
 
@@ -198,34 +197,6 @@ namespace osu.Framework.IO.Stores
         {
             base.Dispose(disposing);
             glyphStores.ForEach(g => g.Dispose());
-        }
-
-        internal sealed class TexturedCharacterGlyph : ITexturedCharacterGlyph
-        {
-            public Texture Texture { get; }
-
-            public float XOffset => glyph.XOffset * ScaleAdjustment;
-            public float YOffset => glyph.YOffset * ScaleAdjustment;
-            public float XAdvance => glyph.XAdvance * ScaleAdjustment;
-            public char Character => glyph.Character;
-            public float Width => Texture.Width * ScaleAdjustment;
-            public float Height => Texture.Height * ScaleAdjustment;
-
-            public readonly float ScaleAdjustment;
-            private readonly CharacterGlyph glyph;
-
-            public TexturedCharacterGlyph(CharacterGlyph glyph, Texture texture, float scaleAdjustment)
-            {
-                this.glyph = glyph;
-                this.ScaleAdjustment = scaleAdjustment;
-
-                Texture = texture;
-            }
-
-            [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public float GetKerning<T>(T lastGlyph)
-                where T : ICharacterGlyph
-                => glyph.GetKerning(lastGlyph) * ScaleAdjustment;
         }
     }
 }
